pyproject.toml
[project]
name = "atlassian-mcp-raw-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.12.2",
    "python-dotenv>=1.1.1",
]
 